make build success on windows
