+++ transport/internet/filelocker.go
@@ -1,0 +1,10 @@
+package internet
+
+import (
+	"os"
+)
+
+type FileLocker struct {
+	path string
+	file *os.File
+}
