+++ transport/internet/filelocker_windows.go
@@ -1,0 +1,9 @@
+package internet
+
+func (fl *FileLocker) Acquire() error {
+	return nil
+}
+
+func (fl *FileLocker) Release() {
+	return
+}
