+++ transport/internet/system_listener.go
@@ -2,12 +2,10 @@ package internet
 
 import (
 	"context"
-	"os"
 	"runtime"
 	"syscall"
 
 	"github.com/pires/go-proxyproto"
-	"golang.org/x/sys/unix"
 	"v2ray.com/core/common/net"
 	"v2ray.com/core/common/session"
 )
@@ -39,36 +37,6 @@ func getControlFunc(ctx context.Context, sockopt *SocketConfig, controllers []co
 				}
 			}
 		})
-	}
-}
-
-type FileLocker struct {
-	path string
-	file *os.File
-}
-
-func (fl *FileLocker) Acquire() error {
-	f, err := os.Create(fl.path)
-	if err != nil {
-		return err
-	}
-	if err := unix.Flock(int(f.Fd()), unix.LOCK_EX); err != nil {
-		f.Close()
-		return newError("failed to lock file: ", fl.path).Base(err)
-	}
-	fl.file = f
-	return nil
-}
-
-func (fl *FileLocker) Release() {
-	if err := unix.Flock(int(fl.file.Fd()), unix.LOCK_UN); err != nil {
-		newError("failed to unlock file: ", fl.path).Base(err).WriteToLog()
-	}
-	if err := fl.file.Close(); err != nil {
-		newError("failed to close file: ", fl.path).Base(err).WriteToLog()
-	}
-	if err := os.Remove(fl.path); err != nil {
-		newError("failed to remove file: ", fl.path).Base(err).WriteToLog()
 	}
 }
